requirement.txt
streamlit==1.29.0
groq==0.7.0
langchain-groq==0.1.3
python-dotenv==1.0.0
pandas==2.2.1
numpy==1.26.3
plotly==5.18.0
pillow==10.2.0
matplotlib==3.8.2
seaborn==0.13.2
requests==2.31.0
pytest==7.4.4
black==24.2.0
flake8==7.0.0